グラフ表示機能の作成 #4
* グラフを作成し、黒石と白石の個数差の推移を表示できるようにした。#4
* GUIのサイズに関するグローバル定数のデータ型をintからdoubleに修正した。
* ResultControllerクラス内で、未使用の引数があるメソッドがある場合は引数を削除した。

+++ src/common/Global.java
@@ -1,7 +1,7 @@
 package common;
 
 public class Global {
-    
+
     /* リバーシに関する定数 */
     /** リバーシ盤の幅（マス） */
     public static final int BOARD_WIDTH = 8;
@@ -11,32 +11,32 @@ public class Global {
 
     /** デフォルトで使用するアルゴリズムの要素番号 */
     public static final int DEFAULT_ALGORITHM = 0;
-    
+
     /* 画面描画のデザインに関する定数 */
     /** ルートペインの横幅 */
-    public static final int ROOT_PANE_WIDTH = 1152;
+    public static final double ROOT_PANE_WIDTH = 1152.0;
 
     /** ルートペインの高さ */
-    public static final int ROOT_PANE_HEIGHT = 864;
+    public static final double ROOT_PANE_HEIGHT = 864.0;
 
     /** 結果表示のタブ内に表示するペインの横幅 */
-    public static final int RESULT_TAB_PANE_WIDTH = 640;
+    public static final double RESULT_TAB_PANE_WIDTH = 640.0;
 
     /** 結果表示のタブ内に表示するペインの高さ */
-    public static final int RESULT_TAB_PANE_HEIGHT = 480;
-    
+    public static final double RESULT_TAB_PANE_HEIGHT = 480.0;
+
     /** アルゴリズム選択のラジオボタンのオブジェクトの幅 */
-    public static final int RADIO_BUTTON_WIDTH = 280;
+    public static final double RADIO_BUTTON_WIDTH = 280.0;
 
     /** アルゴリズム選択のラジオボタンのオブジェクトの高さ */
-    public static final int RADIO_BUTTON_HEIGHT = 40;
-    
+    public static final double RADIO_BUTTON_HEIGHT = 40.0;
+
     /** ゲーム中に表示する、リバーシ盤のマスの大きさ（縦・横同じ） */
-    public static final double GRID_SIZE = 70;
-    
+    public static final double GRID_SIZE = 70.0;
+
     /** 結果画面に表示する、リバーシ盤のマスの大きさ（縦・横同じ） */
-    public static final double GRID_SIZE_RESULT = 50;
-    
+    public static final double GRID_SIZE_RESULT = 50.0;
+
     /* 画面描画のタイマーイベントに関する定数 */
     /** 1秒間に画面描画する回数(FPS) */
     public static final int FPS = 30;
@@ -46,7 +46,7 @@ public class Global {
 
     /** 最初に画面描写してからゲーム開始するまでに処理を待つ時間（ミリ秒） */
     public static final int WAIT_MILLISEC_START = 800;
-    
+
     /** ターン間に操作を待つ時間（ミリ秒） */
     public static final int WAIT_MILLISEC_INTERVAL = 500;
 }
